+++ HV/src/main/java/de/ocplearn/hv/model/Contact.java
@@ -2,7 +2,7 @@ package de.ocplearn.hv.model;
 
 public class Contact implements Comparable<Contact>{
 
-	private long id;
+	private int id;
 	private String sex;
 	private String firstName;
 	private String lastName;
@@ -18,7 +18,7 @@ public class Contact implements Comparable<Contact>{
 		
 		// Artikel zum Thema: https://dzone.com/articles/design-patterns-the-builder-pattern
 		
-		private long id;
+		private int id;
 		private String sex;
 		private String firstName;
 		private String lastName;
@@ -63,7 +63,7 @@ public class Contact implements Comparable<Contact>{
 		/**
 		 * @param id the id to set
 		 */
-		public ContactBuilder setId(long id) {
+		public ContactBuilder setId(int id) {
 			this.id = id;
 			return this;
 		}
